refactor: extract scan, discovery, connectors routes from server.py (Phase 3)
server.py: 1,487 → 399 lines (73% reduction this phase, 87% total).

New route modules:
- routes/scan.py (599 lines) — 16 scan endpoints + SSE stream + helpers
- routes/discovery.py (329 lines) — 4 agent discovery endpoints + lifecycle graph
- routes/connectors.py (225 lines) — connectors, registry, malicious check, scorecard

Also moves _executor ThreadPoolExecutor to pipeline.py (shared scan infra).
All 5,503 tests pass. Test patch targets updated for moved functions.
